Fixed download history page
Automatic polling was reimplemented to manually refreshable feature with
new button. Automatic polling was disabled in order to preserve
functionality of column resizing. Also column Download URL got slider if
the link is longer then specified width and the width of all columns was
adjusted to fit better into the page.